\name{meanC}
\alias{meanC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Calculation of mean and variance

}
\description{Mean and Variance Calculation Using C program in R platform}
\usage{
meanC(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
 \code{x} A vector
}
}
\details{
Mean and Variance Calculation Using C program in R platform
}
\value{
Mean of input vector \code(x)
}

\author{
Nishith Kumar
}

