\name{varC}
\alias{varC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Variance Calculation
}
\description{
Variance calculation Using C In R
}
\usage{
varC(x)
}

\arguments{
  \item{x}{
Vector
}
}
\details{
Variance calculation
}
\value{
Variance
}

\author{
Nishith Kumar
}
